refactor(voice): modularize frontend tools
