+++ .env.example
@@ -9,6 +9,13 @@ DASHSCOPE_API_KEY=
 # QWEN_AUDIO_WEB_SEARCH_MCP_TOKEN=
 # QWEN_AUDIO_WEB_SEARCH_MCP_TOOL=web_search
 
+# 可选前台工具默认按已配置的 Provider 和运行能力自动启用；设为 false 可隐藏。
+# QWEN_AUDIO_SCHEDULE_TOOL_ENABLED=false
+# QWEN_AUDIO_WEB_TOOLS_ENABLED=false
+# QWEN_AUDIO_KNOWLEDGE_TOOL_ENABLED=false
+# QWEN_AUDIO_NOTES_TOOL_ENABLED=false
+# QWEN_AUDIO_RECALL_TOOL_ENABLED=false
+
 # 可选：选择 Gateway 统一使用的 DashScope Realtime 模型；修改后需重启 Gateway。
 # 支持 qwen3.5-omni-flash-realtime、qwen3.5-omni-plus-realtime、
 # qwen-audio-3.0-realtime-flash 和默认的 qwen-audio-3.0-realtime-plus。
