+++ server/src/voice/tools/tool-result.mjs
@@ -1,0 +1,14 @@
+export function toolFailure(errorCode, userMessage, {
+  retryable = false,
+  status = 'failed',
+  ...details
+} = {}) {
+  return {
+    status,
+    error: true,
+    error_code: errorCode,
+    user_message: userMessage,
+    retryable,
+    ...details,
+  }
+}
